+++ src/module/actor/lancer-actor.ts
@@ -158,6 +158,29 @@ export class LancerActor extends Actor {
   }
 
   /**
+   * Returns the current overcharge roll/text
+   * Only applicable for pilots
+   * Overkill for now but there are situations where we'll want this to be configurable
+   */
+  getOverchargeRoll(): string | null {
+    // Function is only applicable to pilots.
+    if (this.data.type !== "pilot") return null;
+
+    const data = this.data as LancerPilotActorData;
+
+    switch (data.data.mech.overcharge_level) {
+      case 1:
+        return "1d3";
+      case 2:
+        return "1d6";
+      case 3:
+        return "1d6+4";
+      default:
+        return "1";
+    }
+  }
+
+  /**
    * Change Class or Tier on a NPC. Recalculates all stats on the NPC.
    * @param newNPCClass Stats object from the new Class.
    */
@@ -304,20 +327,5 @@ export function npc_tier_selector(tier: LancerNPCData["tier"]) {
     <option value="npc-tier-3" ${tier === "npc-tier-3" ? "selected" : ""}>TIER 3</option>
     <option value="npc-tier-custom" ${tier === "npc-tier-custom" ? "selected" : ""}>CUSTOM</option>
   </select>`;
-  return template;
-}
-
-/**
- * Handlebars helper for an overcharge button
- * Currently this is overkill, but eventually we want to support custom overcharge values 
- * @param level Level of overcharge, between 0 (1) and 3 (1d6+4)
- */
-export function overcharge_button(level: number) {
-  let template = 
-   `<div class="overcharge-container">
-      <a class="overcharge-button" style="width:90%;height:90%">
-        1
-      </a>
-    </div>`;
   return template;
 }